Remove temporary compatibility layer
This commit cleans up the temporary compatibility layer
so we now interact directly with the new classes
