+++ polling_stations/apps/data_finder/helpers/__init__.py
@@ -1,5 +1,4 @@
 from .directions import DirectionsHelper
-from .every_election import EveryElectionWrapper
 from .geocoders import (
     PostcodeError,
     geocode,
@@ -14,6 +13,5 @@ __all__ = [
     geocode_point_only,
     geocode,
     get_council,
-    EveryElectionWrapper,
     RoutingHelper,
 ]
